try to split files
