+++ example/integration_test/models_test.dart
@@ -1,0 +1,25 @@
+import 'package:flutter_test/flutter_test.dart';
+import 'package:maplibre/maplibre.dart';
+
+void modelTests() {
+  test('QueriedLayer', () {
+    const o = QueriedLayer(
+      layerId: 'layerId',
+      sourceId: 'sourceId',
+      sourceLayer: 'sourceLayer',
+    );
+    final o2 = QueriedLayer(
+      layerId: o.layerId,
+      sourceId: o.sourceId,
+      sourceLayer: o.sourceLayer,
+    );
+
+    expect(o, equals(o2));
+    expect(o.hashCode, o2.hashCode);
+
+    final oString = o.toString();
+    expect(oString, contains(o.layerId));
+    expect(oString, contains(o.sourceId));
+    expect(oString, contains(o.sourceLayer));
+  });
+}
